MOD: prueba

+++ vtex/orders_api/ORDERS/descomposicion_orders/item.py
@@ -244,7 +244,7 @@ def run():
         "mode": "NULLABLE"
     },{
         "name": "tax",
-        "type": "FLOAT",
+        "type": "INTEGER",
         "mode": "NULLABLE"
     },{
         "name": "priceValidUntil",
